Mejora de schema y show columns

+++ cli.py
@@ -2,6 +2,7 @@ import sys
 import os
 import pathlib
 import shutil
+import json
 from antlr4 import *
 from sqlLexer import sqlLexer
 from sqlParser import sqlParser
@@ -34,34 +35,60 @@ class plsListener(ParseTreeListener):
         except Exception as e:
             print (bdne)
 
+    def exitShow_columns_stmt(self, ctx:sqlParser.Show_columns_stmtContext):
+        try:
+            if bdActual != " ":
+                data = json.load(open(ctx.table_name().getText()+"/schema.json"))
+                for x in data['campos']:
+                    const = ""
+                    for y in data ['constraints']:
+                        if y['columna'] == x['nombre']:
+                            const = const + ", " + y['constraint']
+                    print("     *" + x['nombre'] + " de tipo " + x['tipo'] + " " + const)
+                print(ctx.table_name().getText())
+            else:
+                print(ingresePls)
+        except Exception as e:
+            print (e)
+            print (bdne)
+
 
     def exitCreate_table_stmt(self, ctx:sqlParser.Create_table_stmtContext):
-        if(oie!=" "):
+        if(bdActual!=" "):
+            dict = {'nombre':ctx.table_name().getText(), 'campos':[], 'constraints':[] }
             os.makedirs(ctx.table_name().getText())
-            f=open(ctx.table_name().getText()+"/"+"schema.txt", "w+")
+            #f=open(ctx.table_name().getText()+"/"+"schema.txt", "w+")
             d=open(ctx.table_name().getText()+"/"+"data.txt", "w+")
             d.close()
             for x in range(len(ctx.column_def())):
-                f.write(ctx.column_def()[x].column_name().getText())
-                f.write(":"+ctx.column_def()[x].type_name().getText()+"\n")
+                #f.write(ctx.column_def()[x].column_name().getText())
+                #f.write(":"+ctx.column_def()[x].type_name().getText()+"\n")
+                dict['campos'].append({'nombre':ctx.column_def()[x].column_name().getText(), 'tipo':ctx.column_def()[x].type_name().getText()})
             for x in range(len(ctx.table_constraint())):
                 for y in range(len(ctx.table_constraint()[x].column_name())):
                     print(ctx.table_constraint()[x].column_name()[y].getText())
                     if ctx.table_constraint()[x].K_PRIMARY()!=None:
                         print("PRIMARY")
-                        f.write("PRIMARY_KEY:"+ctx.table_constraint()[x].column_name()[y].getText()+"\n")
+                        #f.write("PRIMARY_KEY:"+ctx.table_constraint()[x].column_name()[y].getText()+"\n")
+                        dict['constraints'].append({'columna':ctx.table_constraint()[x].column_name()[y].getText(), 'constraint':'PRIMARY KEY'})
                     if ctx.table_constraint()[x].K_FOREIGN()!=None:
                         print("FOREIGN")
-                        f.write("FOREIGN_KEY:"+ctx.table_constraint()[x].column_name()[y].getText()+" REFERENCES:"+ctx.table_constraint()[x].foreign_key_clause().foreign_table().getText()+"\n")
+                        #f.write("FOREIGN_KEY:"+ctx.table_constraint()[x].column_name()[y].getText()+" REFERENCES:"+ctx.table_constraint()[x].foreign_key_clause().foreign_table().getText()+"\n")
+                        dict['constraints'].append({'columna':ctx.table_constraint()[x].column_name()[y].getText(), 'constraint':'REFERENCES'+ ctx.table_constraint()[x].foreign_key_clause().foreign_table().getText()})
                     if ctx.table_constraint()[x].K_UNIQUE()!=None:
                         print("UNIQUE")
-                        f.write("UNIQUE:"+ctx.table_constraint()[x].column_name()[y].getText()+"\n")
+                        #f.write("UNIQUE:"+ctx.table_constraint()[x].column_name()[y].getText()+"\n")
+                        dict['constraints'].append({'columna':ctx.table_constraint()[x].column_name()[y].getText(), 'constraint':'UNIQUE'})
                     if ctx.table_constraint()[x].K_CHECK()!=None:
                         print("CHECK")
-                        f.write("CHECK:"+ctx.table_constraint()[x].column_name()[y].getText()+"\n")
-            f.close()
+                        #f.write("CHECK:"+ctx.table_constraint()[x].column_name()[y].getText()+"\n")
+                        dict['constraints'].append({'columna':ctx.table_constraint()[x].column_name()[y].getText(), 'constraint':'CHECK'})
+            #f.close()
+            print (dict)
+            with open(ctx.table_name().getText()+"/"+"schema.json", "w+") as outfile:
+                json.dump(dict, outfile)
         else:
-            print("Not in a DATABASE")
+            print(ingresePls)
 
 
     def exitShow_tables_stmt(self, ctx:sqlParser.Show_tables_stmtContext):
@@ -102,7 +129,13 @@ class plsListener(ParseTreeListener):
 
 
     def exitDrop_table_stmt(self, ctx:sqlParser.Drop_table_stmtContext):
-        shutil.rmtree(ctx.table_name().getText())
+        if bdActual != " ":
+            try:
+                shutil.rmtree(ctx.table_name().getText())
+            except Exception as e:
+                print("Esa tabla no existe Xd")
+        else:
+            print(ingresePls)
 
 
 
@@ -142,7 +175,7 @@ Las construcciones validas para esta gramatica son todas aquellas
 def main(argv):
     while True:
         try:
-            text = input(bdActual + "> ")
+            text = input(" " + bdActual + " > ")
 
             if (text == 'exit'):
                 sys.exit()
